requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-multipart
email-validator
pydantic
pyjwt
pymysql
python-dotenv
cryptography
